pyproject.toml
[project]
name = "wiseagents"
version = "0.0.1"
description = "Multi-agent framework"
authors = []
readme = "README.md"
requires-python = ">=3.8.1,<4.0"

dependencies = [
    "langchain==0.2.16",
    "langchain-community==0.2.16",
    "langchain-huggingface==0.0.3",
    "langchain-openai==0.1.7",
    "langchain-postgres==0.0.9",
    "neo4j==5.21",
    "pydantic==2.7.4",
    "sentence-transformers==3.0.1",
    "streamlit==1.36.0",
    "pyyaml==6.0.1",
    "stomp.py==8.1.2",
    "mkdocstrings-python",
    "mkdocs",
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
    "mkdocs-section-index",
    "mkdocs-include-markdown-plugin",
    "redis",
]

[project.optional-dependencies]
test = [
    "pytest",
]

[pytest]
pythonpath = "src"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
markers = [
    "needsllm: The LMM is too big to run in GitHub Actions, so we need to disable tests requiring it there.",
    "needsllama: Needs llama, and the model is too big for GitHub Actions.",
    "needsqroq: Needs an API key for Groq to be able to run successfully.",
]